requirements.txt
pyyaml
rich
click
requests